[Refactor] Admin::Api::Services::ProxiesControllerTest from Functional to Integration
